removed old fusion and indexing modules + generalized Flask run method
