--- python/smqtk/fusion/__init__.py
@@ -1,13 +1,0 @@
-# -*- coding: utf-8 -*-
-"""
-LICENCE
--------
-Copyright 2015 by Kitware, Inc. All Rights Reserved. Please refer to
-KITWARE_LICENSE.TXT for licensing information, or contact General Counsel,
-Kitware, Inc., 28 Corporate Drive, Clifton Park, NY 12065.
-
-SMQTK subsystem providing the ability to manage and fuse index rankings from
-multiple descriptor-indexer pairings.
-
-"""
-
