+++ python/smqtk/iqr_control/__init__.py
@@ -1,3 +1,2 @@
 from .iqr_session import IqrSession
-from .iqr_session_fusion import IqrSessionFusion
 from .iqr_controller import IqrController
